moved CSS into Sass partials
