+++ packages/frontend/src/components/Counter/counter.scss
@@ -1,0 +1,44 @@
+.section-counter {
+    height: 90vh;
+    position: relative;
+  }
+  
+  .counter__text-box {
+    position: absolute;
+    top: 40%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    text-align: center;
+  }
+  
+  .counter__value {
+    font-size: 12rem;
+  }
+  
+  .counter__button{
+    border: none;
+    cursor: pointer;
+    padding: 1.5rem 4rem;
+    border-radius: 10rem;
+    margin: 0 2rem;
+  }
+  
+  .counter__button:active{
+    transform: translateY(-.1rem);
+    box-shadow: 0 .5rem 1rem rgba(black, 0.2);  
+  }
+  
+  .counter__button:focus{
+    outline: none;
+  }
+  
+  .btn--red{
+    background-color:lightcoral;
+    color: white;
+  }
+  
+  .btn--green{
+    background-color:lightgreen;
+    color: white;
+  }
+
