+++ packages/frontend/src/sass-globals/_base.scss
@@ -1,0 +1,10 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: inherit;
+}
+  
+body {
+    padding: 3rem;
+    box-sizing: border-box;
+}
