+++ packages/frontend/src/sass-globals/_utilities.scss
@@ -1,0 +1,7 @@
+.utility-text-center {
+    text-align: center;
+  }
+  
+.utility-margin-bottom-small {
+    margin-bottom: 0.5rem;
+}
