+++ packages/frontend/src/style.scss
@@ -1,76 +1,9 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: inherit;
-}
+// Import Sass Partials
 
-body {
-  font-family: 'Lato', sans-serif;
-  font-weight: 400;
-  line-height: 1.7;
-  color: #777;
-  /* background-color: orangered; */
-  padding: 3rem;
-  box-sizing: border-box;
-}
+// Global partials
+@import "sass-globals/base";
+@import "sass-globals/typography";
+@import "sass-globals/utilities";
 
-.section-counter {
-  height: 90vh;
-  position: relative;
-}
-
-.counter__text-box {
-  position: absolute;
-  top: 40%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-}
-
-.u-text-center {
-  text-align: center;
-}
-
-.u-margin-bottom-small {
-  margin-bottom: 0.5rem;
-}
-
-.heading-primary {
-  font-size: 3rem;
-  border: 0.1rem solid #777;
-  border-radius: 10px;
-  background: linear-gradient(178deg, lightcoral 50%, lightgreen 50%);
-  -webkit-background-clip: text;
-  color: transparent;
-}
-
-.counter__value {
-  font-size: 12rem;
-}
-
-.counter__button{
-  border: none;
-  cursor: pointer;
-  padding: 1.5rem 4rem;
-  border-radius: 10rem;
-  margin: 0 2rem;
-}
-
-.counter__button:active{
-  transform: translateY(-.1rem);
-  box-shadow: 0 .5rem 1rem rgba(black, 0.2);  
-}
-
-.counter__button:focus{
-  outline: none;
-}
-
-.btn--red{
-  background-color:lightcoral;
-  color: white;
-}
-
-.btn--green{
-  background-color:lightgreen;
-  color: white;
-}
+// Component styles
+@import "components/Counter/counter.scss";
